requirements.txt
genson==1.2.2
jsonschema==4.20.0
tiktoken==0.9.0
pydantic==2.11.5
openai==1.57.0
pandas==2.1.4
python-dotenv==1.0.1
sentence-transformers==4.0.2
torch==2.7.1
torchvision==0.22.1
spacy==3.7.2
numpy==2.3.0
scikit-learn==1.7.0
nltk==3.9
docx2txt==0.8
pdfminer.six==20231228
python-dateutil==2.8.2
datasets==2.21.0
tqdm==4.66.5
boto3==1.35.0
urllib3==2.2.3
tabulate==0.9.0